db+sqlite+pg: split ListAddresses into its own file
Move ListAddresses and IterAddresses into addressstore_listaddresses.go in
each backend. ListAddresses is a paginated list read (analogous to ListTxns /
ListUTXOs), so per the wallet/internal/db README it stays a direct backend
implementation rather than an ops interface.
