+++ wallet/internal/db/pg/addresses.go
@@ -2,57 +2,14 @@ package pg
 
 import (
 	"context"
-	"database/sql"
 	"fmt"
-	"iter"
 	"time"
 
 	"github.com/btcsuite/btcwallet/wallet/internal/db"
-	"github.com/btcsuite/btcwallet/wallet/internal/db/page"
 	"github.com/btcsuite/btcwallet/wallet/internal/sql/pg/sqlc"
 )
 
 var _ db.AddressStore = (*Store)(nil)
-
-// ListAddresses returns a page of addresses matching the given query.
-func (s *Store) ListAddresses(ctx context.Context,
-	query db.ListAddressesQuery) (page.Result[db.AddressInfo, uint32], error) {
-
-	if query.Page.Limit() == 0 {
-		return page.Result[db.AddressInfo, uint32]{}, db.ErrInvalidPageLimit
-	}
-
-	var items []db.AddressInfo
-
-	err := s.execRead(ctx, func(q *sqlc.Queries) error {
-		var err error
-
-		items, err = listAddressesByAccount(ctx, q, query)
-
-		return err
-	})
-	if err != nil {
-		return page.Result[db.AddressInfo, uint32]{}, err
-	}
-
-	result := page.BuildResult(
-		query.Page, items,
-		func(item db.AddressInfo) uint32 {
-			return item.ID
-		},
-	)
-
-	return result, nil
-}
-
-// IterAddresses returns an iterator over paginated address results.
-func (s *Store) IterAddresses(ctx context.Context,
-	query db.ListAddressesQuery) iter.Seq2[db.AddressInfo, error] {
-
-	return page.Iter(
-		ctx, query, s.ListAddresses, db.NextListAddressesQuery,
-	)
-}
 
 // NewDerivedAddress creates a new address for a given account and key
 // scope.
@@ -334,102 +291,4 @@ func applyAddressAccountMetadata(info *db.AddressInfo,
 		row.MasterFingerprint, row.Purpose, row.CoinType,
 		!row.IsDerived,
 	)
-}
-
-// listAddressesByAccount lists addresses filtered by wallet ID, key scope,
-// and account name, with pagination support.
-func listAddressesByAccount(ctx context.Context, q *sqlc.Queries,
-	query db.ListAddressesQuery) ([]db.AddressInfo, error) {
-
-	if query.Scope == nil && query.AccountName == nil {
-		return listRawImportedAddresses(ctx, q, query)
-	}
-
-	if query.Scope == nil || query.AccountName == nil {
-		return nil, db.ErrInvalidListAddressesQuery
-	}
-
-	rows, err := q.ListAddressesByAccount(
-		ctx, buildAddressPageParams(query),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("list addresses by account: %w", err)
-	}
-
-	items := make([]db.AddressInfo, len(rows))
-	for i, row := range rows {
-		item, err := addressRowToInfo(row)
-		if err != nil {
-			return nil,
-				fmt.Errorf("list addresses by account: map address row: %w",
-					err)
-		}
-
-		items[i] = *item
-	}
-
-	return items, nil
-}
-
-// listRawImportedAddresses lists raw imported addresses for the imported alias.
-func listRawImportedAddresses(ctx context.Context, q *sqlc.Queries,
-	query db.ListAddressesQuery) ([]db.AddressInfo, error) {
-
-	rows, err := q.ListRawImportedAddresses(
-		ctx, sqlc.ListRawImportedAddressesParams{
-			WalletID:  int64(query.WalletID),
-			PageLimit: int64(query.Page.Limit()) + 1,
-			CursorID:  rawAddressCursor(query),
-		},
-	)
-	if err != nil {
-		return nil, fmt.Errorf("list raw imported addresses: %w", err)
-	}
-
-	items := make([]db.AddressInfo, len(rows))
-	for i, row := range rows {
-		item, err := addressRowToInfo(row)
-		if err != nil {
-			return nil, fmt.Errorf("list raw imported addresses: %w", err)
-		}
-
-		items[i] = *item
-	}
-
-	return items, nil
-}
-
-// rawAddressCursor converts an optional page cursor into a nullable sqlc value.
-func rawAddressCursor(q db.ListAddressesQuery) sql.NullInt64 {
-	if q.Page.After == nil {
-		return sql.NullInt64{}
-	}
-
-	return sql.NullInt64{
-		Int64: int64(*q.Page.After),
-		Valid: true,
-	}
-}
-
-// buildAddressPageParams translates a ListAddresses query to
-// ListAddressesByAccount parameters, handling pagination cursors.
-func buildAddressPageParams(
-	q db.ListAddressesQuery) sqlc.ListAddressesByAccountParams {
-
-	params := sqlc.ListAddressesByAccountParams{
-		WalletID:    int64(q.WalletID),
-		Purpose:     int64(q.Scope.Purpose),
-		CoinType:    int64(q.Scope.Coin),
-		AccountName: *q.AccountName,
-		PageLimit:   int64(q.Page.Limit()) + 1,
-	}
-
-	if q.Page.After != nil {
-		params.CursorID = sql.NullInt64{
-			Int64: int64(*q.Page.After),
-			Valid: true,
-		}
-	}
-
-	return params
 }
